#1268 Step3 Customer Execute Payment

+++ order/src/main/java/com/yas/order/model/Order.java
@@ -52,10 +52,10 @@ public class Order extends AbstractAuditEntity {
     private String note;
 
     @Column(name = "total_tax")
-    private float tax;
+    private BigDecimal tax;
 
     @Column(name = "total_discount_amount")
-    private float discount;
+    private BigDecimal discount;
 
     private int numberItem;
